model/requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
docopt==0.6.2
docopt-ng==0.6.3
filelock==3.14.0
flatbuffers==24.3.25
flwr==1.5.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
hydra-core==1.3.2
idna==3.7
iterators==0.0.2
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
loguru==0.6.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mosaik==3.3.3
mosaik-api-v3==3.0.12
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==2.8.8
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==3.20.3
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.6.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vessim==0.8.0
Werkzeug==3.0.3
wrapt==1.16.0